About page created and styled
