--- client/js/screens/AboutUs/index.js
@@ -1,3 +1,0 @@
-import AboutUs from './AboutUs';
-
-export default AboutUs;
